verified MxTimer::MxTimer() and moved some booleans to MxBool
It was assumed Windows' BOOL would be the same as LEGO Island's MxBool but this
is actually not true. Windows' BOOL is an `int` (32-bit signed) and Island's
MxBool is an `unsigned char`.

It's also theorized that Mindscape used 0xFF for `TRUE` rather than 0x1. This
could explain a lot of the code discrepancies.